requirements.txt
tensorflow==2.15.0
flask==3.0.0
pillow==10.1.0
numpy==1.24.3
scikit-learn==1.3.2
python-dotenv==1.0.0
gunicorn==21.2.0 